First library separation
