+++ Paths_syfco.hs
@@ -13,15 +13,29 @@
 --
 -----------------------------------------------------------------------------
 
+{-# OPTIONS_HADDOCK
+
+    hide
+  , prune
+  , ignore-exports
+
+  #-}
+
+-----------------------------------------------------------------------------
+
 module Paths_syfco
   ( version,
   ) where
 
 -----------------------------------------------------------------------------
 
-import Data.Version (Version(..))
+import Data.Version
+  ( Version(..)
+  )
 
 -----------------------------------------------------------------------------
+
+-- | Fallback version set to zero.
 
 version
   :: Version
