--- src/Writer/Modes.hs
@@ -1,12 +1,0 @@
-module Writer.Modes where
-
----
-
-data WriteMode =
-    Pretty
-  | Fully
-  deriving (Eq, Show)
-           
----
-
-
