--- src/Writer/Partition.hs
@@ -1,13 +1,0 @@
-module Writer.Partition where
-
----
-
-partitionLtl
-  :: [String] -> [String] -> String
-
-partitionLtl is os =
-  ".inputs" ++ concatMap (' ' :) is ++ "\n" ++
-  ".outputs" ++ concatMap (' ' :) os ++ "\n"
-  
----  
-
